requirements.txt
future
paho-mqtt
python-aqi
APScheduler
Pillow
ffmpeg-python
homeassistant-mqtt-binding
